refactor(web-app): rename layer.ts to entities.layer.ts
Align with the entities.* naming convention used by other entity
files (e.g. entities.event.ts, entities.observation.ts).
