+++ web-app/src/app/admin/admin-event/event-details/event-details.component.ts
@@ -10,7 +10,7 @@ import { NgForm } from '@angular/forms';
 import { ActivatedRoute, Router } from '@angular/router';
 import { CdkDragDrop, moveItemInArray } from '@angular/cdk/drag-drop';
 
-import { Layer } from 'mage-web-app/entities/layer/layer';
+import { Layer } from 'mage-web-app/entities/layer/entities.layer';
 import { Event as MageEvent } from 'mage-web-app/entities/event/entities.event';
 import { AdminBreadcrumb } from '../../admin-breadcrumb/admin-breadcrumb.model';
 import { AdminBreadcrumbService } from '../../admin-breadcrumb/admin-breadcrumb.service';
@@ -25,7 +25,7 @@ import {
 import { DeleteEventComponent } from '../delete-event/delete-event.component';
 import { CreateEventDialogComponent } from '../create-event/create-event.component';
 import { UploadFormDialogComponent } from '../upload-form/upload-form.component';
-import { layerIconName } from '../../../entities/layer/layer';
+import { layerIconName } from '../../../entities/layer/entities.layer';
 
 interface ExtendedEvent extends MageEvent {
   complete?: boolean;
